Move the Minecraft transformers to classes
