+++ src/mc1.15.2/java/me/dreamhopping/pml/launch/PMLEntryPoint.java
@@ -10,6 +10,7 @@ import me.dreamhopping.pml.impl.MinecraftImpl;
 import me.dreamhopping.pml.transformers.ClientPlayNetworkHandlerTransformer;
 import me.dreamhopping.pml.transformers.ClientPlayerEntityTransformer;
 import me.dreamhopping.pml.transformers.InGameHudTransformer;
+import me.dreamhopping.pml.transformers.MinecraftClientTransformer;
 import net.minecraft.client.main.Main;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.tree.*;
@@ -23,37 +24,7 @@ public class PMLEntryPoint {
         classLoader.registerTransformer(new InGameHudTransformer());
         classLoader.registerTransformer(new ClientPlayerEntityTransformer());
         classLoader.registerTransformer(new ClientPlayNetworkHandlerTransformer());
-
-        // Transform the window title
-        classLoader.registerTransformer(new RuntimeTransformer() {
-            @Override
-            public boolean willTransform(String name) {
-                return name.equals("net/minecraft/client/MinecraftClient");
-            }
-
-            @Override
-            public ClassNode transform(ClassNode classNode) {
-                for (MethodNode methodNode : classNode.methods) {
-                    if (methodNode.name.equals("getWindowTitle")) {
-                        InsnList list = new InsnList();
-                        list.add(new LdcInsnNode(" | PufferfishModLoader (dev/1.0)"));
-                        list.add(new MethodInsnNode(Opcodes.INVOKEVIRTUAL, "java/lang/StringBuilder", "append", "(Ljava/lang/String;)Ljava/lang/StringBuilder;"));
-
-                        // Find the "toString" call and insert before that
-                        for (AbstractInsnNode node = methodNode.instructions.getLast(); node != null; node = node.getPrevious()) {
-                            if (node.getOpcode() == Opcodes.INVOKEVIRTUAL) {
-                                MethodInsnNode castedNode = (MethodInsnNode) node;
-                                if (castedNode.owner.equals("java/lang/StringBuilder") && castedNode.name.equals("toString") && castedNode.desc.equals("()Ljava/lang/String;")) {
-                                    methodNode.instructions.insertBefore(castedNode, list);
-                                }
-                            }
-                        }
-                    }
-                }
-
-                return classNode;
-            }
-        });
+        classLoader.registerTransformer(new MinecraftClientTransformer());
 
         Minecraft.setInstance(new MinecraftImpl());
