+++ src/mc1.16.4/java/me/dreamhopping/pml/transformers/MinecraftClientTransformer.java
@@ -1,0 +1,35 @@
+package me.dreamhopping.pml.transformers;
+
+import me.dreamhopping.pml.mods.launch.loader.RuntimeTransformer;
+import org.objectweb.asm.Opcodes;
+import org.objectweb.asm.tree.*;
+
+public class MinecraftClientTransformer implements RuntimeTransformer {
+    public boolean willTransform(String name) {
+        return name.equals("net/minecraft/client/MinecraftClient");
+    }
+
+    public ClassNode transform(ClassNode classNode) {
+        for (MethodNode methodNode : classNode.methods) {
+            if (methodNode.name.equals("getWindowTitle")) {
+                InsnList list = new InsnList();
+                list.add(new LdcInsnNode(" | PufferfishModLoader (dev/1.0)"));
+                list.add(new MethodInsnNode(Opcodes.INVOKEVIRTUAL, "java/lang/StringBuilder", "append", "(Ljava/lang/String;)Ljava/lang/StringBuilder;"));
+
+                // Find the "toString" call and insert before that
+                for (AbstractInsnNode node = methodNode.instructions.getLast(); node != null; node = node.getPrevious()) {
+                    if (node.getOpcode() == Opcodes.INVOKEVIRTUAL) {
+                        MethodInsnNode castedNode = (MethodInsnNode) node;
+                        if (castedNode.owner.equals("java/lang/StringBuilder") && castedNode.name.equals("toString") && castedNode.desc.equals("()Ljava/lang/String;")) {
+                            methodNode.instructions.insertBefore(castedNode, list);
+                        }
+                    }
+                }
+
+                break;
+            }
+        }
+
+        return classNode;
+    }
+}
